feature: added a home webpage with search functionality at http://localhost:8080/home.
The home page lists five random artists, tracks and genres as music suggestions while laso featuring a search field where the user can search for tracks.

The search page features a back button which take the user back to the Home-page and it also lists the tracks from the search results. If there was no input it will prompt that the search field was empty.